fix: create scan submitted twice
