+++ app/(app)/scan/licence-plate-camera.tsx
@@ -2,28 +2,24 @@ import { router, usePathname } from 'expo-router'
 import { useCallback, useEffect, useRef, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { View } from 'react-native'
-import { useSafeAreaInsets } from 'react-native-safe-area-context'
 import { Camera } from 'react-native-vision-camera'
 
 import { TorchState } from '@/components/camera/FlashlightBottomSheetAttachment'
+import { LicencePlateCameraBackground } from '@/components/camera/LicencePlateCameraBackground'
 import LicencePlateCameraBottomSheet from '@/components/camera/LicencePlateCameraBottomSheet'
 import OcrCamera from '@/components/camera/OcrCamera'
+import { HomeButton } from '@/components/navigation/HomeButton'
 import ScreenView from '@/components/screen-layout/ScreenView'
 import DismissKeyboard from '@/components/shared/DissmissKeyboard'
-import IconButton from '@/components/shared/IconButton'
 import { ScanResultEnum } from '@/modules/backend/openapi-generated'
-import {
-  CROPPED_AREA_HEIGHT,
-  HEADER_WITH_PADDING,
-  useScanLicencePlate,
-} from '@/modules/camera/hooks/useScanLicencePlate'
+import { useScanLicencePlate } from '@/modules/camera/hooks/useScanLicencePlate'
 import { TextData } from '@/modules/camera/types'
 import { useOffenceStoreContext } from '@/state/OffenceStore/useOffenceStoreContext'
 import { useSetOffenceState } from '@/state/OffenceStore/useSetOffenceState'
 import { addTextToImage } from '@/utils/addTextToImage'
-import { cn } from '@/utils/cn'
 
-let plates: string[] = []
+const MAX_PLATE_HISTORY = 5
+const REQUIRED_SCANS = 2
 
 const LicencePlateCameraComp = () => {
   const { t } = useTranslation()
@@ -31,7 +27,7 @@ const LicencePlateCameraComp = () => {
   const [torch, setTorch] = useState<TorchState>('off')
   const [isManual, setIsManual] = useState(false)
 
-  const { top } = useSafeAreaInsets()
+  const plateHistoryRef = useRef<string[]>([])
 
   const pathname = usePathname()
 
@@ -54,7 +50,6 @@ const LicencePlateCameraComp = () => {
 
       setOffenceState({ photos: [imageWithTimestampUri] })
     } catch {
-      // TODO: Debug why this occurs twice... error happens when the camera is closed and picture is taken, needs further investigation
       console.log('error')
     }
   }, [ref, setOffenceState])
@@ -66,17 +61,21 @@ const LicencePlateCameraComp = () => {
       const ecv = scanLicencePlate(frame)
 
       if (ecv) {
-        if (!plates.includes(ecv)) {
-          plates.push(ecv)
-          if (plates.length > 5) plates.shift()
+        const plateHistory = plateHistoryRef.current
+        plateHistory.push(ecv)
 
-          return
+        if (plateHistory.length > MAX_PLATE_HISTORY) {
+          plateHistory.shift()
         }
 
-        setOffenceState({ ecv })
-        takeLicencePlatePicture()
+        const scannedSamePlates = plateHistory.filter((plate) => plate === ecv).length
 
-        await checkEcv(ecv, isManual)
+        if (scannedSamePlates === REQUIRED_SCANS) {
+          setOffenceState({ ecv })
+          takeLicencePlatePicture()
+
+          await checkEcv(ecv, isManual)
+        }
       }
     },
     [generatedEcv, checkEcv, isManual, scanLicencePlate, setOffenceState, takeLicencePlatePicture],
@@ -98,7 +97,7 @@ const LicencePlateCameraComp = () => {
 
   const onChangeLicencePlate = useCallback(
     (ecv: string) => {
-      plates = []
+      plateHistoryRef.current = []
 
       setIsManual(!!ecv)
       setOffenceState({ scanResult: undefined, ecv })
@@ -120,38 +119,19 @@ const LicencePlateCameraComp = () => {
     }
   }, [scanResult, onChangeLicencePlate, pathname])
 
-  const backgroundClassName = cn('items-center bg-dark/75', {
-    'bg-green/75': scanResult === ScanResultEnum.NoViolation,
-    'bg-negative/75': scanResult === ScanResultEnum.PaasParkingViolation,
-    'bg-warning/75': scanResult === ScanResultEnum.PaasParkingViolationDuplicity,
-  })
-
   return (
     <DismissKeyboard>
       <ScreenView
         title={t('scanLicencePlate.title')}
         options={{
-          headerRight: () => (
-            <IconButton
-              name="home"
-              accessibilityLabel={t('offenceResult.home')}
-              onPress={() => router.navigate('/')}
-            />
-          ),
+          headerRight: () => <HomeButton />,
         }}
         className="h-full"
       >
         <View className="relative">
           <OcrCamera ref={ref} torch={torch} onFrameCapture={onFrameCapture} />
 
-          <View className="absolute h-full w-full">
-            <View
-              style={{ paddingTop: top, height: HEADER_WITH_PADDING }}
-              className={cn('justify-start', backgroundClassName)}
-            />
-            <View style={{ height: CROPPED_AREA_HEIGHT }} className="items-center" />
-            <View className={cn('flex-1 ', backgroundClassName)} />
-          </View>
+          <LicencePlateCameraBackground />
         </View>
 
         <LicencePlateCameraBottomSheet
