+++ components/navigation/HomeButton.tsx
@@ -1,0 +1,16 @@
+import { router } from 'expo-router'
+import { useTranslation } from 'react-i18next'
+
+import IconButton from '@/components/shared/IconButton'
+
+export const HomeButton = () => {
+  const { t } = useTranslation()
+
+  return (
+    <IconButton
+      name="home"
+      accessibilityLabel={t('offenceResult.home')}
+      onPress={() => router.navigate('/')}
+    />
+  )
+}
